third_party/WebKit/LayoutTests/fast/events/wheelevent-ctrl-expected.txt

TOP



Scroll mouse wheel over here



END
And scroll the document here
Tests that wheel events with the ctrl modifier are handled properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test mousewheel events over scrollable div
With ctrl modifier set
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is true
PASS testDiv.scrollTop is 0
Without ctrl
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is false
PASS testDiv.scrollTop is deltaY

Test mousewheel events over the document
With ctrl modifier set
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is true
PASS document.body.scrollTop is 0
Now without ctrl
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is false
PASS window.scrollY is deltaY
PASS successfullyParsed is true

TEST COMPLETE
